refactor: split four files sitting at their length caps
AppStore.swift and SessionCommands.swift were at 1000 and 997 lines against a
1000-line limit, ControlDispatcherTests.swift and AppStoreTests.swift at 1969
and 1981 against 2000, and the Session command struct at 795 against an
800-line type limit. Any addition to them fails make lint.

pure movement, no behavior change: the flag and background-watermark mutators
go to AppStore+SessionState.swift, the status/restore/flag CLI subcommands to
SessionMetadataCommands.swift, and the matching test blocks to
ControlDispatcherSessionMetadataTests.swift and AppStoreTreeProjectionTests.swift.
same 2796 tests pass, in 114 suites rather than 112.
